requirement.txt
streamlit
pandas
numpy
Pillow
torch
torchvision
faiss-cpu
ftfy
regex
tqdm
git+https://github.com/openai/CLIP.git
requests
# Stable Diffusion
diffusers
transformers
accelerate
safetensors
scipy
# Optional
sentencepiece
